+++ test/fixtures/yard/nested_array.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class NestedArray
+  # @param [Array<Array<String>>] array
+  # @return [Array<Array<String>>]
+  def nested_array_method(array)
+    array
+  end
+end