+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add packwerk to dependencies https://github.com/euglena1215/packwerk-yard/pull/6
 - Introduce sorbet https://github.com/euglena1215/packwerk-yard/pull/7
+- Support nested array & nested hash https://github.com/euglena1215/packwerk-yard/pull/8
 
 ## [0.1.0] - 2024-01-07